+++ src/pages/main/Album.tsx
@@ -1,14 +1,14 @@
-import EmotionTag from "../components/common/EmotionTag";
-import SearchBar from "../components/pages/album/SearchBar";
+import EmotionTag from "../../components/common/EmotionTag";
+import SearchBar from "../../components/pages/album/SearchBar";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Appbar from "../components/common/Appbar";
-import { getSearchedDiaries } from "../api/api";
-import ThumbnailGrid from "../components/common/ThumbnailGrid";
-import ROUTE_PATH from "../constants/ROUTE_PATH";
-import useInView from "../hooks/useInView";
-import useScrollPosition from "../hooks/useScrollPosition";
-import { createQueryParams } from "../utils/util";
+import Appbar from "../../components/common/Appbar";
+import { getSearchedDiaries } from "../../api/api";
+import ThumbnailGrid from "../../components/common/ThumbnailGrid";
+import ROUTE_PATH from "../../constants/ROUTE_PATH";
+import useInView from "../../hooks/useInView";
+import useScrollPosition from "../../hooks/useScrollPosition";
+import { createQueryParams } from "../../utils/util";
 import { useInfiniteQuery } from "@tanstack/react-query";
 
 const tags = ["기쁨", "슬픔", "분노", "공포", "혐오", "수치", "놀람", "궁금", "무난"];
