+++ src/pages/main/Explore.tsx
@@ -1,10 +1,10 @@
-import ThumbnailGrid from "../components/common/ThumbnailGrid";
-import Appbar from "../components/common/Appbar";
+import ThumbnailGrid from "../../components/common/ThumbnailGrid";
+import Appbar from "../../components/common/Appbar";
 import { useState, useRef, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import ROUTE_PATH from "../constants/ROUTE_PATH";
-import { getExploreDiaries } from "../api/api";
-import useInView from "../hooks/useInView";
+import ROUTE_PATH from "../../constants/ROUTE_PATH";
+import { getExploreDiaries } from "../../api/api";
+import useInView from "../../hooks/useInView";
 import { useInfiniteQuery } from "@tanstack/react-query";
 
 const buttonPositions = {
